Move refusals to its own example
